Move _mock-bitcore from src
